+++ sdk/resourcemanager/containerservicefleet/armcontainerservicefleet/models.go
@@ -10,6 +10,83 @@ package armcontainerservicefleet
 
 import "time"
 
+// APIServerAccessProfile - Access profile for the Fleet hub API server.
+type APIServerAccessProfile struct {
+	// Whether to create the Fleet hub as a private cluster or not.
+	EnablePrivateCluster *bool
+
+	// Whether to enable apiserver vnet integration for the Fleet hub or not.
+	EnableVnetIntegration *bool
+
+	// The subnet to be used when apiserver vnet integration is enabled. It is required when creating a new Fleet with BYO vnet.
+	SubnetID *string
+}
+
+// AgentProfile - Agent profile for the Fleet hub.
+type AgentProfile struct {
+	// The ID of the subnet which the Fleet hub node will join on startup. If this is not specified, a vnet and subnet will be
+	// generated and used.
+	SubnetID *string
+
+	// The virtual machine size of the Fleet hub.
+	VMSize *string
+}
+
+// AutoUpgradeProfile - The AutoUpgradeProfile resource.
+type AutoUpgradeProfile struct {
+	// The resource-specific properties for this resource.
+	Properties *AutoUpgradeProfileProperties
+
+	// READ-ONLY; If eTag is provided in the response body, it may also be provided as a header per the normal etag convention.
+	// Entity tags are used for comparing two or more entities from the same requested resource.
+	// HTTP/1.1 uses entity tags in the etag (section 14.19), If-Match (section 14.24), If-None-Match (section 14.26), and If-Range
+	// (section 14.27) header fields.
+	ETag *string
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string
+
+	// READ-ONLY; The name of the resource
+	Name *string
+
+	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
+	SystemData *SystemData
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string
+}
+
+// AutoUpgradeProfileListResult - The response of a AutoUpgradeProfile list operation.
+type AutoUpgradeProfileListResult struct {
+	// REQUIRED; The AutoUpgradeProfile items on this page
+	Value []*AutoUpgradeProfile
+
+	// The link to the next page of items
+	NextLink *string
+}
+
+// AutoUpgradeProfileProperties - The properties of the AutoUpgradeProfile.
+type AutoUpgradeProfileProperties struct {
+	// REQUIRED; Configures how auto-upgrade will be run.
+	Channel *UpgradeChannel
+
+	// If set to False: the auto upgrade has effect - target managed clusters will be upgraded on schedule. If set to True: the
+	// auto upgrade has no effect - no upgrade will be run on the target managed
+	// clusters. This is a boolean and not an enum because enabled/disabled are all available states of the auto upgrade profile.
+	// By default, this is set to False.
+	Disabled *bool
+
+	// The node image upgrade to be applied to the target clusters in auto upgrade.
+	NodeImageSelection *NodeImageSelection
+
+	// The resource id of the UpdateStrategy resource to reference. If not specified, the auto upgrade will run on all clusters
+	// which are members of the fleet.
+	UpdateStrategyID *string
+
+	// READ-ONLY; The provisioning state of the AutoUpgradeProfile resource.
+	ProvisioningState *AutoUpgradeProfileProvisioningState
+}
+
 // ErrorAdditionalInfo - The resource management error additional info.
 type ErrorAdditionalInfo struct {
 	// READ-ONLY; The additional info.
@@ -83,6 +160,27 @@ type FleetCredentialResult struct {
 type FleetCredentialResults struct {
 	// READ-ONLY; Array of base64-encoded Kubernetes configuration files.
 	Kubeconfigs []*FleetCredentialResult
+}
+
+// FleetHubProfile - The FleetHubProfile configures the fleet hub.
+type FleetHubProfile struct {
+	// The access profile for the Fleet hub API server.
+	APIServerAccessProfile *APIServerAccessProfile
+
+	// The agent profile for the Fleet hub.
+	AgentProfile *AgentProfile
+
+	// DNS prefix used to create the FQDN for the Fleet hub.
+	DNSPrefix *string
+
+	// READ-ONLY; The FQDN of the Fleet hub.
+	Fqdn *string
+
+	// READ-ONLY; The Kubernetes version of the Fleet hub.
+	KubernetesVersion *string
+
+	// READ-ONLY; The Azure Portal FQDN of the Fleet hub.
+	PortalFqdn *string
 }
 
 // FleetListResult - The response of a Fleet list operation.
@@ -163,6 +261,9 @@ type FleetPatch struct {
 
 // FleetProperties - Fleet properties.
 type FleetProperties struct {
+	// The FleetHubProfile configures the Fleet's hub.
+	HubProfile *FleetHubProfile
+
 	// READ-ONLY; The status of the last operation.
 	ProvisioningState *FleetProvisioningState
 }
@@ -331,6 +432,22 @@ type OperationListResult struct {
 
 	// READ-ONLY; List of operations supported by the resource provider
 	Value []*Operation
+}
+
+// SkipProperties - The properties of a skip operation containing multiple skip requests.
+type SkipProperties struct {
+	// REQUIRED; The targets to skip.
+	Targets []*SkipTarget
+}
+
+// SkipTarget - The definition of a single skip request.
+type SkipTarget struct {
+	// REQUIRED; The skip target's name. To skip a member/group/stage, use the member/group/stage's name; Tp skip an after stage
+	// wait, use the parent stage's name.
+	Name *string
+
+	// REQUIRED; The skip target type.
+	Type *TargetType
 }
 
 // SystemData - Metadata pertaining to creation and last modification of the resource.